pyproject.toml
[project]
name = "portscanner-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-nmap>=0.7.1",
]
